# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ErrorJSON:
    properties:
      error:
        type: string
    type: object
...
info:
  contact: {}
  description: 'REST API: health, chain status, counter contract calls, bank ledger
    (PostgreSQL indexer + optional The Graph subgraph).'
  title: go-chain API
  version: 0.1.0
paths:
...
  /api/code-pulse/campaigns:
    get:
      parameters:
      - description: Filter by state
        in: query
        name: state
        type: string
      - description: Filter by proposal_id
        in: query
        name: proposal_id
        type: integer
      - description: Filter by organizer address
        in: query
        name: organizer
        type: string
      - description: Filter by developer (joined)
        in: query
        name: developer
        type: string
      - description: Filter by contributor (joined)
        in: query
        name: contributor
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: 'Sort: launched_at_desc (default), deadline_at_asc'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List campaigns
      tags:
      - code-pulse
  /api/code-pulse/campaigns/{campaignId}:
    get:
      parameters:
      - description: Campaign ID
        in: path
        name: campaignId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorJSON'
      summary: Campaign detail
      tags:
      - code-pulse
  /api/code-pulse/campaigns/{campaignId}/contributions:
    get:
      parameters:
      - description: Campaign ID
        in: path
        name: campaignId
        required: true
        type: integer
      - description: Filter by contributor address
        in: query
        name: contributor
        type: string
      - description: 'Sort: amount_desc (default), latest'
        in: query
        name: sort
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Campaign contributions
      tags:
      - code-pulse
  /api/code-pulse/campaigns/{campaignId}/timeline:
    get:
      parameters:
      - description: Campaign ID
        in: path
        name: campaignId
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Campaign timeline
      tags:
      - code-pulse
...
  /api/code-pulse/contributors/{address}/dashboard:
    get:
      parameters:
      - description: Wallet address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Contributor dashboard
      tags:
      - code-pulse
  /api/code-pulse/developers/{address}/dashboard:
    get:
      parameters:
      - description: Wallet address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Developer dashboard
      tags:
      - code-pulse
  /api/code-pulse/initiators/{address}/dashboard:
    get:
      parameters:
      - description: Wallet address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Initiator dashboard
      tags:
      - code-pulse
...
swagger: "2.0"